Add a bunch of basic functionality unit tests.  The naming is self-explanatory--for example, name_preorder_register_update_cantforce will create a namespace, preorder a name, register it, update it, try to update it with a different key, and verify that the different key can't force the update.
Each test allows the programmer to define a set of wallets filled with
btc in advance, create a blockchain by describing the order of
operations on a block-by-block basis, and then query the resulting
blockstore database to verify that each operation was processed
correctly.  The 'testlib.py' file serves as a common interface between
blockstore's API and the test logic--it will be stable for all tests, so
blockstore can evolve independently.