Remember which block ID contained the consensus hash that accompanied a NAME_TRANSFER operation.  This is not a consensus-breaking change, but any database that does not incorporate this information on NAME_TRANSFER will be unverifiable.
The block ID is used to later look up the consensus hash that
accompanied the NAME_TRANSFER, when verifying the database or performing
SNV.  This is necessary because due to a quirk in the original
Blockstore prototype, NAME_TRANSFER did not save its consensus hash, and
our test framework accidentally masked this behavior (by sheer
coincidence, it could always select the "right" consensus hash when
reconstructing a NAME_TRANSFER operation from a name's history when
running in the test environment).  This has since been caught, and
integration tests have been written to watch for this bug.