Add info about snap install and continuous delivery
Because only an upstream maintainer would be able to use the simple continuous delivery infrastructure that we provide in https://build.snapcraft.io, I had to build my own pipeline for continuous delivery.

I let my imagination run wild, to play with the boundaries of our integrations, and ended up with a git branch that mutates itself :) Any reviews on that Travis-CI job will also be appreciated.

We now have v0.14.4.1 in the beta channel for all the supported architectures, and a process to update the snap in the edge channel every day without manual intervention. So I will start the call for testing with the Ubuntu community as soon as you approve this branch and process.
If things work nicely, I hope to release the next tag to the candidate channel, and make the following tag our first stable release.

Any suggestions are welcome.