test: add tests for traits with duplicate methods
In Clarity1, these are allowed. When used, only the last instance of the
method with the same name is visible to users of the trait. In Clarity2,
defining a trait like this will cause an error, but a Clarity2 contract
may use a trait defined in Clarity1, and it should exhibit the same
behavior as a Clarity1 user (last method only is usable).