feat: after each tx, re-walk the list of txs
This matches the behavior of the previous mempool walk algorithm and
improves on the ordering of txs. Without this change, if the fee rate
for nonce 2 for account A is greater than the fee rate for nonce 1, then
on the first pass, first nonce 2 is skipped, then nonce 1 is included,
then other transactions with lower fee rates are included before nonce 2
is included. With this change, the walk starts back at the beginning of
the fee rate-sorted list after each considered transaction.