feat: improvements to mempool iteration
These changes take lessons learned from experiments by @gregorycoppola
and myself, as well as feedback on various related PRs. It uses the
following techniques to improve the speed and scalability of the mempool
walk.
* Uses rusqlite's `Rows` iterator to read one row at a time
* Caches the nonces in memory to avoid repeated lookups
* Restarts search from the highest fee-rate transactions after every
  executed transaction
  * Caches potential transactions in memory to retry on next pass

With this implementation, miners can reliably fill a block in <30s,
regardless of how large the mempool gets.