feat: change behavior of sqlite nonce cache
Originally, this cache was only used after the in-memory cache is full.
Now, it is used more like a typical L2 cache, where it is always in sync
with the L1 cache, i.e. when the L1 cache is updated, so is the L2 cache.
Since this L2 cache is preserved across iterations, it improves the
speed in the typical running behavior.