test: fix tests' reliance on incorrect nonce
These tests were dependent on the old implementations increment of
nonces when a transaction was skipped. The correct response is to only
increment the nonces when the transaction is successfully included.
Therefore these tests need to simulate a success event in order to get
the expected behavior to be tested.