fix: type-check when type cannot be determined
Addresses Immunefi #13019. In this expression, `(index-of (list) none)`
there is no way to determine the complete type of the list. We can only
infer that it is an `optional`, but we cannot figure out the inner type.
This results in the expected type being `NoType` which causes
`admits_type` to panic. This commit contains two different solutions
to this problem that I would like to discuss.