fix: the heaviest affirmation map calculation is different in epoch 2.05 -- it's the same as the sortition affirmation map.
Normally, there's no difference because the heaviest block-commit by BTC spend (the 2.1 rule) is the same as the Stacks block-commit with the most confirmations (the 2.05 rule). However, in the current testnet, this is not the case for reward cycle #360. This means that once the 2.1 rules take effect, the node will get stuck -- it will think that the stacks and sortition affirmation maps are not in sync with the heaviest affirmation map, and the node will never sync as a result.

This fix makes it so the heaviest affirmation map -- and by extension, the canonical affirmation map -- is calculated from the sortition affirmation map for reward cycle in epoch 2.05 and earlier.