fix: revert all regex changes in the v1 parser
With the addition of the v2 parser, some of the regex expressions were
moved up to representations.rs so that they could be shared and reused.
This has proven to be a problem, because the strange behavior of the v1
parser was not preserved (e.g. allowing long contract names in
contract-call parameters, but not in contracts).

Fixes: #3470