docs: consistent markdown formatting
The main goal of this commit is to resolve the issues where a newline
was intended, but because Markdown formatting uses two newlines to
separate paragraphs, these lines were being merged. This was especially
apparent in the lists of possible error codes, which I've now
reformatted to use bulleted lists.