chore: output logs to stdout instead of stderr.
Rationale: most tools expect programs to output to `stderr`.
`stderr` should be used when there's an actual error (e.g. the program
fails to start because of incorrect options). The current behavior
causes surprises and frustrations (e.g. trying to `tee` the output of
stacks-node to a file only to realize the file's empty).