fix: replace async code in event dispatcher
The reason behind this is that with the async code we had, we were
unable to successfully implement a timeout, so when there was a network
glitch, it would get stuck at `TcpStream::connect`, and crash after 60
minutes. This new implementation ensures that all parts of the event
dispatcher networking code are able to fail with a timeout.