fix: unused imports
Added a cfg to warn about unused imports in `net/mod.rs`, which exposed a few unused imports that I've removed.

I've also updated other parts of the code that had imports that were only used for testing. In those cases, I've updated the import to only be used in `cfg(test)`.