chore: fix two flaky tests
Fixes:

* signer::v0::locally_rejected_blocks_overriden_by_global_acceptance
* signer::v0::reorg_locally_accepted_blocks_across_tenures_succeeds

These tests used to depend on the `blocks_mined` counter and then immediately check a `v2/info` assertion -- this is a race condition: the stacks-node may not have processed the mined block yet. This caused test flake in CI (but usually not in local runs where machines are fast enough to never experience this condition).