feat(wasm): use only one `Engine` per global context
Instantiating a `wasmtime::Engine` is an expensive operation, so it is
best to do it only once for the duration of a `GlobalContext`. Cloning
an engine is, however, not an expensive operation (just an `Arc::clone`)
and we use it to avoid referring to the global context already referred
to by the instantiated `ClarityWasmContext`.

See-also: https://github.com/stacks-network/clarity-wasm/issues/468