test: add `tenure_extend_after_2_bad_commits`
This test checks if a miner is able to extend its tenure twice in a row
due to bad block commits. Currently, this test fails because when the
miner's original winning tenure is 2 behind, then that miner is no
longer able to write to the StackerDB. This problem needs to be further
discussed as to what the best solution is.