feat: stop waiting for signatures if the chain tip advances
This could happen at the beginning of a tenure, if the miner for tenure
T+1 proposes its first block before it sees the last block from tenure
T, so the block it proposed is reorging that last block. Once it sees
this block, it should immediately stop waiting for signatures and
instead build a new block off of the new tip.

Fixes: #5751