feat: add `tenure_extend_buffer_secs` miner config option
This allows the miner to specify a buffer (in seconds) that it should
wait after it has reached a time when it thinks 70% of signers should
approve a time-based tenure extend. The goal is to provide a little
buffer for clock skew, since we have seen, live on mainnet, cases where
<70% of signers approve a tenure-extend.

This is item 2 in #5757.