fix: miner should not count repeat rejections
Before this fix, when the miner is waiting for responses to its block
proposal, if it received a block rejection more than once, the second
and subsequent rejections would incorrectly count towards the 30%
rejection threshold. This could cause the miner to prematurely treat the
block as rejected, causing it to build and propose a new block. This
change fixes that so that it only counts the response from each signer
once.