feat: on timeout, re-propose the same block
When a miner times out waiting for signatures, instead of proposing a
new block, it should only re-propose the same block. Proposing a new
block is guaranteed to fail because signers that approved the original
block will reject any new block at the same height.

This implements the miner side of #5856. A change is still needed on the
signer side to allow a signer to accept a block that it previously
rejected.