fix: remove the `candidate_cache` during mempool iteration
Now that the db query is taking the nonces into account, it no longer
makes sense to cache these candidate transactions. On the next query,
their nonces should be properly set in the db, so that will get ordered
correctly. This cache would only make performance worse in that case.