fix: ignore `wait_on_interim_blocks`
The wait that this introduced was there just to ensure that the chains
coordinator had time to set the flag to pause the miner before it
attempted to mine again. This is now handled instead by waiting for the
last block to be processed before proceeding.

To force a longer wait time between blocks, there is already the
`min_time_between_blocks_ms` option for this.