feat: wait for miner to be unblocked between blocks
In addition to waiting for the last block it mined to be processed, now
the miner will also wait to be unblocked. This ensures that the miner
does not start mining before it is unblocked by the chains coordinator
thread.

This solves flakiness in `tests::signer::v0::verify_mempool_caches`.