fix: do not break UTXO chain
Bad timing between the Stacks node's processing and the Bitcoin network
including a block commit can cause a miner to submit a block commit
which breaks the UTXO chain, severly hurting its chances of winning
blocks. It is better to just miss this commit than to break the UTXO
chain.