feat: add more prometheus metrics for mempool statistics
Miners (and mock-miners) can report how many transactions are (could be)
included in a block and the reason that they stopped building a block.
This information is very helpful to understand whether the block was
full, the miner ran out of time, the miner was preempted, or there were
no transactions available to mine.