fix: get last locally or globally accepted block in tenure
In `get_tenure_last_block_info`, we should return the last block that we
know about in a tenure, whether that is a locally accepted block or a
globally accepted block. Previously, this would only return a locally
accepted block, which would cause signers to incorrectly reject blocks
if they did not have a locally accepted block which was higher than the
globally accepted block.