test: fix flakiness in tests that use `TEST_MINE_STALL`
In CI, we would sometimes see the case where the test set this flag, to
stall the miner, then submits a transaction, expecting it not to
immediately be included in a block, but because of the timing, the miner
was already in a loop collecting transactions for the next block and
this transaction could be picked up before the miner reaches the stall.

This change creates a function for tests to use to set the stall and
then wait for the miner to reach the stall before returning, removing
this source of flakiness.