refactor: use `Vec<Value>` for allowances
This replaces a `HashSet<String>` which was using the serialized value
for quick lookups, but added the extra overhead on setup and use of
serializing the values. With this new design, the value is used
directly, with no new allocations, and then a linear search is required
through the `Vec`, but in the common case, the number of entries will be
very low.