refactor: make `EventObserver` private
The only reason it wasn't private (anymore) was the regression test
that's being deleted here, added in 9ad9ead83518c4d4c8bbabf8cc8980cca639353c.
By deleting it, I can remove the `pub`s.

Since deleting a regression test is probably controversial, let me
justify why I think it's both warranted and acceptable.

The reason I want to do it is that for https://github.com/stacks-network/stacks-core/issues/6543
I'll have to mess around with the implementation of the event delivery
pipeline, and if there's less public surface to worry about, I'll have
to make fewer compromises.

As for "acceptable", this integration test injects a MITM into the event
pipeline to modify event payloads (this is why it needed to make the
`EventObserver` public), because the bug that was fixed with #5860 was
that the miner's key was incorrectly grabbed from the event data.

But as of 4b5019fa91c03922d1227a3de458443c6a21d4b2, the signer's event
handling code no longer passes around the miner key at all, making it
extremely unlikely for a regression to ever occur that would be caught
by this test.