fix: do not double hash in `secp256r1-verify`
This is a consensus breaking change and will need to go out in a
hard-fork with a new version of Clarity (Clarity5).
In the existing implementation, the code was incorrectly hashing the
passed in `message-hash`. Unfortunately, the test code was also
double-hashing on the signing side as well, so this was not caught. The
bug was discovered by @eric-stacks when attempting to generate
signatures externally (e.g. with WebAuthn) and verify them in Clarity
code.