refactor: remove `db_path` from `EventObserver`
This is the final step to making `EventObserver` a "dumb" data object
that simply reflects the configuration of the registered observer.

Most of this change is straightforward. There is one thing that's worth
a mention, and that's the `ProposalCallbackHandler`, a concept that was
added in https://github.com/stacks-network/stacks-core/pull/4228.
Because block proposal validation happens on a dedicated thread, those
validation events are handled a little specially.

Since `send_payload` now needs a dispatcher instance as the source of
the `db_path`, the list of observers alone is no longer enough.

For now I've elected to simply add the DB path to the
ProposalCallbackHandler and keep a non-method version of `send_payload`
around that gets the DB path passed in. Eventually I may have to clone
the dispatcher, but for now this approach had the smallest blast radius.

There wasn't any test coverage for this special behavior (at least not
on the event dispatcher side -- there are [some tests](https://github.com/stacks-network/stacks-core/blob/45381558f5e79554fc6e7930b0c4091e739e4992/stackslib/src/net/api/tests/postblock_proposal.rs#L234)
on the API side of things), so I added that as well.