refactor: remove the ability to use event dispatcher without DB
This ability was only used during tests; in actual production, there is
always a DB.

Not having to support the DB-less use is going to make it a lot cleaner
to use the DB as a queue (see #6543).

This commit also makes the method names and order of operations a little
less confusing -- we had `send_payload()`, `send_payload_directly()`,
`send_payload_given_db_path()` (ok that one was my fault),
and `send_payload_with_bytes()`.

Now we have this instead:

```
dispatch_to_observer()
-> get_payload_bytes()
-> save_to_db()
-> make_http_request_and_delete_from_db()
   -> make_http_request()
   -> delete_from_db()
```